+++ app/domain/secrets/secret_types/ephemeral_secret_type.rb
@@ -4,6 +4,8 @@ module Secrets
       def input_validation(params)
 
       end
+
+
     end
   end
 end